refactor (activity): 샘플데이터 생성 mainActivity에서 제거
